requirements.txt
boto3>=1.40.0
botocore>=1.40.0
kubernetes>=33.1.0
pyyaml>=6.0
prometheus-client>=0.22.1
